adventurer/sign/door indicator interaction system
